separate build context functions
