+++ src/buildFunctions.js
@@ -62,6 +62,166 @@ function getRootIndexRanges( geo ) {
 
 }
 
+
+// computes the union of the bounds of all of the given triangles and puts the resulting box in target. If
+// centroidTarget is provided then a bounding box is computed for the centroids of the triangles, as well.
+// These are computed together to avoid redundant accesses to bounds array.
+function getBounds( bounds, offset, count, target, centroidTarget = null ) {
+
+	let minx = Infinity;
+	let miny = Infinity;
+	let minz = Infinity;
+	let maxx = - Infinity;
+	let maxy = - Infinity;
+	let maxz = - Infinity;
+
+	let cminx = Infinity;
+	let cminy = Infinity;
+	let cminz = Infinity;
+	let cmaxx = - Infinity;
+	let cmaxy = - Infinity;
+	let cmaxz = - Infinity;
+
+	const includeCentroid = centroidTarget !== null;
+	for ( let i = offset * 6, end = ( offset + count ) * 6; i < end; i += 6 ) {
+
+		const cx = bounds[ i + 0 ];
+		const hx = bounds[ i + 1 ];
+		const lx = cx - hx;
+		const rx = cx + hx;
+		if ( lx < minx ) minx = lx;
+		if ( rx > maxx ) maxx = rx;
+		if ( includeCentroid && cx < cminx ) cminx = cx;
+		if ( includeCentroid && cx > cmaxx ) cmaxx = cx;
+
+		const cy = bounds[ i + 2 ];
+		const hy = bounds[ i + 3 ];
+		const ly = cy - hy;
+		const ry = cy + hy;
+		if ( ly < miny ) miny = ly;
+		if ( ry > maxy ) maxy = ry;
+		if ( includeCentroid && cy < cminy ) cminy = cy;
+		if ( includeCentroid && cy > cmaxy ) cmaxy = cy;
+
+		const cz = bounds[ i + 4 ];
+		const hz = bounds[ i + 5 ];
+		const lz = cz - hz;
+		const rz = cz + hz;
+		if ( lz < minz ) minz = lz;
+		if ( rz > maxz ) maxz = rz;
+		if ( includeCentroid && cz < cminz ) cminz = cz;
+		if ( includeCentroid && cz > cmaxz ) cmaxz = cz;
+
+	}
+
+	target[ 0 ] = minx;
+	target[ 1 ] = miny;
+	target[ 2 ] = minz;
+
+	target[ 3 ] = maxx;
+	target[ 4 ] = maxy;
+	target[ 5 ] = maxz;
+
+	if ( includeCentroid ) {
+
+		centroidTarget[ 0 ] = cminx;
+		centroidTarget[ 1 ] = cminy;
+		centroidTarget[ 2 ] = cminz;
+
+		centroidTarget[ 3 ] = cmaxx;
+		centroidTarget[ 4 ] = cmaxy;
+		centroidTarget[ 5 ] = cmaxz;
+
+	}
+
+}
+
+// A stand alone function for retrieving the centroid bounds.
+function getCentroidBounds( bounds, offset, count, centroidTarget ) {
+
+	let cminx = Infinity;
+	let cminy = Infinity;
+	let cminz = Infinity;
+	let cmaxx = - Infinity;
+	let cmaxy = - Infinity;
+	let cmaxz = - Infinity;
+
+	for ( let i = offset * 6, end = ( offset + count ) * 6; i < end; i += 6 ) {
+
+		const cx = bounds[ i + 0 ];
+		if ( cx < cminx ) cminx = cx;
+		if ( cx > cmaxx ) cmaxx = cx;
+
+		const cy = bounds[ i + 2 ];
+		if ( cy < cminy ) cminy = cy;
+		if ( cy > cmaxy ) cmaxy = cy;
+
+		const cz = bounds[ i + 4 ];
+		if ( cz < cminz ) cminz = cz;
+		if ( cz > cmaxz ) cmaxz = cz;
+
+	}
+
+	centroidTarget[ 0 ] = cminx;
+	centroidTarget[ 1 ] = cminy;
+	centroidTarget[ 2 ] = cminz;
+
+	centroidTarget[ 3 ] = cmaxx;
+	centroidTarget[ 4 ] = cmaxy;
+	centroidTarget[ 5 ] = cmaxz;
+
+}
+
+// returns the average coordinate on the specified axis of the all the provided triangles
+export function getAverage( bounds, offset, count, axis ) {
+
+	let avg = 0;
+	for ( let i = offset, end = offset + count; i < end; i ++ ) {
+
+		avg += bounds[ i * 6 + axis * 2 ];
+
+	}
+
+	return avg / count;
+
+}
+
+// precomputes the bounding box for each triangle; required for quickly calculating tree splits.
+// result is an array of size tris.length * 6 where triangle i maps to a
+// [x_center, x_delta, y_center, y_delta, z_center, z_delta] tuple starting at index i * 6,
+// representing the center and half-extent in each dimension of triangle i
+export function computeBounds( geo ) {
+
+	const verts = geo.attributes.position.array;
+	const index = geo.index.array;
+	const triCount = index.length / 3;
+	const bounds = new Float32Array( triCount * 6 );
+
+	for ( let tri = 0; tri < triCount; tri ++ ) {
+
+		const ai = index[ 3 * tri + 0 ] * 3;
+		const bi = index[ 3 * tri + 1 ] * 3;
+		const ci = index[ 3 * tri + 2 ] * 3;
+
+		for ( let el = 0; el < 3; el ++ ) {
+
+			const a = verts[ ai + el ];
+			const b = verts[ bi + el ];
+			const c = verts[ ci + el ];
+			const min = Math.min( a, b, c );
+			const max = Math.max( a, b, c );
+			const halfExtents = ( max - min ) / 2;
+			bounds[ tri * 6 + el * 2 + 0 ] = min + halfExtents;
+			bounds[ tri * 6 + el * 2 + 1 ] = halfExtents;
+
+		}
+
+	}
+
+	return bounds;
+
+}
+
 export function buildTree( geo, options ) {
 
 	ensureIndex( geo );
@@ -115,7 +275,7 @@ export function buildTree( geo, options ) {
 			const left = node.left = new MeshBVHNode();
 			const lstart = offset, lcount = splitOffset - offset;
 			left.boundingData = new Float32Array( 6 );
-			ctx.getBounds( lstart, lcount, left.boundingData, cacheCentroidBounds );
+			getBounds( ctx.bounds, lstart, lcount, left.boundingData, cacheCentroidBounds );
 
 			splitNode( left, lstart, lcount, cacheCentroidBounds, depth + 1 );
 
@@ -123,7 +283,7 @@ export function buildTree( geo, options ) {
 			const right = node.right = new MeshBVHNode();
 			const rstart = splitOffset, rcount = count - lcount;
 			right.boundingData = new Float32Array( 6 );
-			ctx.getBounds( rstart, rcount, right.boundingData, cacheCentroidBounds );
+			getBounds( ctx.bounds, rstart, rcount, right.boundingData, cacheCentroidBounds );
 
 			splitNode( right, rstart, rcount, cacheCentroidBounds, depth + 1 );
 
@@ -144,12 +304,12 @@ export function buildTree( geo, options ) {
 		if ( geo.boundingBox != null ) {
 
 			root.boundingData = boxToArray( geo.boundingBox );
-			ctx.getCentroidBounds( range.offset, range.count, cacheCentroidBounds );
+			getCentroidBounds( ctx.bounds, range.offset, range.count, cacheCentroidBounds );
 
 		} else {
 
 			root.boundingData = new Float32Array( 6 );
-			ctx.getBounds( range.offset, range.count, root.boundingData, cacheCentroidBounds );
+			getBounds( ctx.bounds, range.offset, range.count, root.boundingData, cacheCentroidBounds );
 
 		}
 
@@ -162,7 +322,7 @@ export function buildTree( geo, options ) {
 
 			const root = new MeshBVHNode();
 			root.boundingData = new Float32Array( 6 );
-			ctx.getBounds( range.offset, range.count, root.boundingData, cacheCentroidBounds );
+			getBounds( ctx.bounds, range.offset, range.count, root.boundingData, cacheCentroidBounds );
 
 			splitNode( root, range.offset, range.count, cacheCentroidBounds );
 			roots.push( root );
